0210admin

+++ teamprojectsoomin/jasmine/views.py
@@ -21,7 +21,7 @@ class MainView:
         return render(request, 'jasmine/login.html', context)
 
     def logout(request):
-        if request.session['suser'] != None:
+        if request.session['suser']!= None:
             del request.session['suser'];
         return render(request,'jasmine/home.html')
 
@@ -31,13 +31,20 @@ class MainView:
         try:
             user = UserDb().selectone(id);
             if pwd == user.userpwd:
-                logger.debug(id)
-                request.session['suser'] = id;
-                request.session['susernum'] = user.usernum;
-                context = {
-                  'section':'jasmine/loginok.html',
-                    'loginuser':user
-                };
+                if user.userid == 'admin':
+                    request.session['suser'] = id;
+                    context = {
+                        'section': 'jasmine/loginok.html',
+                        'loginuser': user
+                    };
+                else:
+                    logger.debug(id)
+                    request.session['suser'] = id;
+                    request.session['susernum'] = user.usernum;
+                    context = {
+                        'section': 'jasmine/loginok.html',
+                        'loginuser': user
+                    };
             else:
                 raise Exception;
         except:
@@ -68,6 +75,10 @@ class MainView:
                 'error': ErrorCode.e0001
             };
         return render(request, 'jasmine/home.html', context);
+
+
+
+
 
 def mypage(request):
     usernum = request.GET['usernum'];
@@ -120,8 +131,15 @@ def userupdateimpl(request):
 def userdelete(request):
     id = request.GET['id'];
     try:
-        UserDb().delete(id);
-        request.session['suser'] = None;
+        if id == 'admin':
+            context={
+                'section': 'jasmine/error.html',
+                'error': ErrorCode.e0004
+            }
+            return render(request, 'admin/manage.html', context)
+        else:
+            UserDb().delete(id);
+            request.session['suser'] = None;
     except:
         context = {
             'section': 'jasmine/error.html',
@@ -215,6 +233,19 @@ class admin:
                 'error': ErrorCode.e0011
             };
         return render(request, 'jasmine/admin/adminpage.html', context)
+
+def manage(request):
+    return render(request, 'admin/manage.html')
+
+def userlist(request):
+    ruserlist = UserDb().select();
+    context = {
+        'section':'admin/userlist.html',
+        'userlist' : ruserlist
+    };
+    return render(request, 'admin/manage.html',context)
+
+
 
 
 def map(request):
